requirements.txt
# Deep Learning & Vision
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.38.2
accelerate>=0.30.1
captum>=0.6.0

# Data Processing
datasets>=2.17.1
evaluate>=0.4.1
rouge_score>=0.1.2
nltk>=3.8.1
pillow>=10.2.0
scikit-learn>=1.4.0
pandas>=2.2.1
numpy>=1.26.0

# Visualization & UI
matplotlib>=3.8.3
seaborn>=0.13.2
gradio>=4.19.0

# LLM Integration
google-generativeai>=0.4.0

# Utility
tqdm>=4.66.2
pyyaml>=6.0.1
jupyter>=1.0.0
ipywidgets>=8.1.2
